[trace][rust] Add TraceFuture for tracing futures
TraceFuture adds a duration event when a future is created, dropped, and
every time it's polled. The duration events are connected by flow
events. The "created" duration event unfortunately doesn't encompass the
creation of the wrapped future since the wrapped future must be
constructed before the TraceFuture.

The purpose of the TraceFuture is to aid in tracking the execution of
async rust code:
 - A future passed to Task::spawn may be first polled on a
   different thread from the one it was created on.
 - A future may be polled multiple times across multiple threads.
 - A future may be dropped before becoming Ready.

The main advantage of TraceFuture over async events is that TraceFuture
shows the synchronous steps of an asynchronous task. Async events are
rendered separately from synchronous duration events which makes it
difficult to associate duration events with async events if multiple
async events are overlapping.

Change-Id: I23273be63dd5654b4d295504b1d51b0b8c0fc92a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934058
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>