[devfs] Remove multiplexing from base path
Removes the multiplexing at the base devfs path to just connect to the
device protocol. Keeps the node as a default when things are running
with dfv1 since there are a lot of fvm tests that use devfs paths with
fds.

Change-Id: Ief087ede5dae6d3a694f5d8d5aba79ab9961a64d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934807
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Christopher Anderson <cja@google.com>