[qemu] Use an SMBIOS v2.1 entry point when booting x86 guests
As of QEMU version 8.1.5, the default SMBIOS entry point type is the
v3/64-bit kind instead of the v2.1/32-bit kind. We only support the
latter at the moment.

This QEMU patch introduces the (undocumented??) pc-specific property of
`smbios-entry-point-type` that allows us to override this default:
https://patchwork.kernel.org/project/qemu-devel/patch/20211026151100.1691925-4-ehabkost@redhat.com/

We set this to unblock a future roll to (or past) QEMU version 8.1.5. We
might consider supporting SMBIOS v3 in the future, but no need to block
the roll on that now.

This property was tested locally with both v8.0.3 (the current version)
and v8.1.5: we successfully booted with an SMBIOS v2.1 entry point in
each case.

Fixed: 134894
Change-Id: I237c5ccf90ce8c4dd9a77ea9622942239ed5b4b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/936612
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Owners-Override: Adam Barth <abarth@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>