[superproject] Roll llvm-project/libc [libc] Compile the GPU functions with '-fconvergent-functions' (#70229)
Summary:
This patch simply adds the `-fconvergent-functions` flag to the GPU
compilation. This is in relation to the behaviour of SIMT
architectures under divergence. With the flag, we assume every function
is convergent by default and rely on the compiler's divergence analysis
to transform it if possible.

Fixes: https://github.com/llvm/llvm-project/issues/63853
GitOrigin-RevId: bde11aaf8562d7296b87b278103014b9d513437a
Original-Revision: 37d1e79d9dfbbffdabf540eef62987bba477219c
Roller-URL: https://ci.chromium.org/b/8766198525117424161
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0304e8f09c3ec5eff251aea564f47134f8b1c06d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937789