[kernel][vm] Avoid excessive scanning
Changes the way the access_scan_period is calculated and used to avoid
redundant scans. The intention of the regular access scanner is that
there is at most access_scan_period between scans, but the existing
implementation had a scenario where if did a routine scan at time A,
and then waited till time A+period, if a scan from the page queues
were to happen just moments before (i.e. A+period-delta) then we would
still rescan.

This change corrects this issue and ensures that the periodic scan
only occurs if nothing else triggered a scan for at least the period.

Change-Id: Ifdba8f1ad70ef0e4072a6406f651db2a10fa743b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/937853
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>