[display] Remove id from sysmem debug name
The id is a separate field, and by not including the id in the name we
get the desired porperty that a sysmem(1) name not vary per instance.
In sysmem2 we'll separate name and client_code_id to make it ok to vary
per instance in the name as long as a client_code_id is also set that
doesn't vary per instance.

BUG: b/284073556
Change-Id: Id18e479a0d0e73df87d700993f6a08806100d248
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938443
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>