[ld] Support IE model TLS access
This adds the support for the IE dynamic relocs and tests them.
This includes tracking DF_STATIC_TLS and insisting that the IE
reloc types are only used in a module that set it.  In the
startup dynamic linker, it's actually no problem to just let
any module use these relocs since by definition they are all in
the IE set.  But it's the static linker's responsibility to
identify a IE-using module with DF_STATIC_TLS so that dlopen
can refuse to load it, so it makes sense to insist that the
static linker did the right thing even in the startup context.

Bug: 128502
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I916eadb31070ef65a8ba524a32e36b24bc368263
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/934913
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>