[storage][rust] Fix uses of format!
The argument to anyhow::Context::context will always be constructed even
if the Result is Ok. If the context needs to be formatted then
Context::with_context should be used to only construct the String when
the Result is an Err.

anyhow! and bail! take format arguments so format! isn't necessary.

Change-Id: Ied6a5be3eaeffa5be194ec1ebae6f7e0a9f33018
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939022
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>