[netstack] Populate socket inspect data for test
Formerly, the ArchiveReader would return an empty DiagnosticsHierarchy
if a snapshot was taken for a selector which did not correspond to
Inspect data. After a recent change, it now returns an
Option<DiagnosticsHierarchy>. That change revealed that the
netstack-sys-integration-test did not retrieve Socket Inspect data in
the relevant tests. Create a UDP socket to ensure that such data exists,
and then remove a special case that had been added to account for the
missing Socket data. While I'm here, remove an unnecessary reference
operator and add a comment documenting a subtlety of one of the test
helpers.

Tested: fx test netstack-sys-integration-test
Fixed: 135964
Change-Id: I6efe92d831ad361e860046c7c52cc1cf525e79a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939896
Commit-Queue: Amanda Tait <atait@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>