[bazel_sdk] allow fuchsia_cc_driver to specify binary name
Adds an attribute to fuchsia_cc_driver which allows users to specify the
name of the built driver. For now we keep the default as lib<name>.so
which is different from in-tree until we can migrate our OOT repos to
match the default that we have in-tree.

BUG: b/305314198
Change-Id: Ic34cb87036559c1285b28f319f607d4ece5e8fa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939989
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>