[ui] Route GraphicalPresenter to session
fuchsia.session.scene.Manager is deprectaed and we intend to remove it
eventually. Making this change now allows us to avoid adding it to the
partner SDK, for products which currently need it and are moving OOT.

Remove references to an obsolete bug which said we should remove the
GraphicalPresenter routes. It is no longer relevant.

Part of "Refactoring Session and Window Manager Components" migration

Change-Id: I7c746cdccae44714c9c2f469aa8de60de8f9d5a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940112
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: David Worsham <dworsham@google.com>