[testsharder] Fix computation of duration of multiplied shards.
We recalculate the shard timeout when creating multiplied shards
because if runs are not specified, then we need to fill up the shard up
to the target duration with the multiplied tests and then calculate the
timeout based on the number of runs we determine. However, in the case
where runs are already specified, the duration of running all multiplied
tests may exceed the target duration. In this case, we should just use
the already calculated timeout from sharding the multiplied tests the
first time since no new test runs will be added.

MULTIPLY: `[
  {
    "max_matches": 500,
    "name": "starnix_gvisor_tests",
    "total_runs": 3
  }
]`
Run-All-Tests: true
Change-Id: I7ccab6913b364804454ceaf2b5c56bf72da331e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939879
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>