[debugger] Add OS info to connection.
Exposes the OS type in the IPC reply message from the debug agent, and
exposes this in various places.

Adds the platform and CPU architecture to the "status" output.

Updates the processing of the HelloReply message to allow versioning as
long as the version is set for the "version 0" deserialization. This
required changes on both the debug_agent and zxdb side to properly allow
versioning on both ends.

Change-Id: I699b81700dd1af5a74bc0d538e6bc31200c78a28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939878
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>