[debugger] Handle overlapping ELF libraries.
When searching for loaded ELF libraries, GetElfModulesForProcess() would
avoid adding regions that overlapped with a previous ELF module (as
explained in the long comment in that function).

But this logic didn't account for modules found using the dl_debug_addr
method earlier in the code, and the "end" address of that module would
never be accounted for. This is a problem only when the mapping found
using the dl_debug_addr (the main exe) is the first mapping in memory
(normally not the case on Fuchsia) which is why we never noticed.

Change-Id: I44c1578bdc809ea90afebc567da08d02ce26f3bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933195
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Adam Barth <abarth@google.com>