[superproject] Roll llvm-project/libc [amdgpu][openmp] Treat missing TIMESTAMP_FREQUENCY as non-fatal (#70987)
If you build with dynamic_hsa, the symbol is known and compilation
succeeds. If you then run with a slightly older libhsa, this argument is
not recognised and an error returned. I'd rather the program runs with a
misleading omp wtime than refuses to run at all.

GitOrigin-RevId: 6d6ae5cb6501f939228fe0985948ee49f84d710d
Original-Revision: d42e77987aedb3bcc239599d9c41ba23eb7a9a14
Roller-URL: https://ci.chromium.org/b/8765576892815501777
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Id7f22a4c695c3e18ca6b563990f257301d413b15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940441