[time][test] Add a realm proxy implementation
The realm proxy instantiates a test realm for Timekeeper,
almost identical to that constructed in timekeeper_integration,
except the API boundary is not rust code, but rather FIDL.

This will allow us to swap different flavors of identical tests
in and out at will.

While this commit is a small milestone, I expect that there
will be further changes in this code, once I get a better
idea what is needed for integration testing.

Bug: b/306024715
Change-Id: I9f3a4e2962c202c5e262ecf60edc82d57fd99b1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939980
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>