[kernel][riscv] skip starting the boot processor when parsing topology
Additionally test against the known boot hart when skipping cpus to try
to start as a secondary. This makes the code somewhat more reslient to a
topology tree not having the PRIMARY flag set, which isn't of very much
use anyway on riscv, since the boot hart is known.

Change-Id: I178e7a793d02a530fa290b235385c3b7a1cc6f31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940162
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>