[swd-e2e-tests] Max size should be on system image size
The max size we are testing with is the system image or update images
size, which should exclude the update package size, since that cannot
get garbage collected during an OTA.

This also restructures how the size is calculated by moving the
computation into the Package types to deduplicate code.

As part of this, it changes the argument from `-max-ota-size` to
`-max-system-image-size`.

Change-Id: Ia03c373198676feefd78cbb0b3fd43451099a908
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940444
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ben Keller <galbanum@google.com>