[devshell] Use mac-arm64 binaries when on arm64 mac
We have brought up all but a handful of (mostly unused) tools that
worked on amd64 macs with arm64 prebuilts.  Let's allow them to be used
directly so people with such machines no longer have to run builds under
Rosetta.

This is also expected to fix the incremental build failures on the m1
bot in FYI, since now it should be able to find the correct
fuchsia-vendored-python python3 binary on PATH, since it'll look in the
same place that we've actually populated the prebuilts.

Bug: b/283475308
Change-Id: I205842211209215f4418b31f6d6fb3868e57cb6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941173
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>