[elf_runner] Use exception handling dedicated job
...so that a component may be able to take its immediate parent's job's
exception channel for exception handling.

The ELF runner attempts to associate exceptions made by the component or
its child jobs/processes to the component itself. To do this, the ELF
runner takes the component's newly created job's exception channel (ELF
runner creates a new job per component to hold the component process) to
receive all (unhandled) exceptions made in that job. For mor details,
see `//src/sys/lib/elf_runner/src/crash_handler.rs`.

The problem with the approach taken by the ELF runner is that it prevents
anything else that is interested in receiving exceptions from a group of
processes underneath (and alongside) the component process because only
a single exception channel may exist for a job[1].

This change solves the above problem by optionally (through a program flag
`job_with_excp_channel`) that creates an extra job between the job the ELF
runner uses for exception tracking and the component process so that the
process will be a child of a job who's exception channel may be safely taken.

This is needed to that later changes that try to use unified address
spaces in Starnix through the use of the `usercopy` crate can install
a single `Usercopy` instance for the whole Starnix process instead of
one per Starnix-managed Linux process.

[1]: https://fuchsia.dev/fuchsia-src/reference/syscalls/task_create_exception_channel?hl=en

Bug: b/276973344, b/309126987
Change-Id: Ic5e4f85f95cb49231839a36ea0bd1adbfce07f8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941052
Reviewed-by: Gary Bressler <geb@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>