[netcfg] Name already exists from naming policy
DynamicNameCompositionRules will need to know information about existing
interfaces to know whether the name needs to be altered, such as in the
case of finding a unique number to normalize a MAC address.

Accessing the interface names during name generation will make it
easier to determine whether an interface with the same name already
exists.

Test added to ensure that an error is surfaced when an interface with
the same name already exists in the Config.

Bug: 135087
Bug: 136013
Bug: b/309010767
Change-Id: I5a9d715d090af142268c3cbf5202ccf3ef2ec2fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939876
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>