[config-netstack] Isolate unused endpoints in test
All the endpoints in the configurable-netstack-test are on the same
broadcast domain, which means in the connect_to_server test, all the
endpoints that are used in the other test cases are also receiving the
traffic from the first test's client and server. This causes a lot of
unnecessary processing and log noise, so instead place them in an
isolated network (that sees no traffic).

Bug: 136090
Multiply: configurable-netstack-test
Change-Id: Iefed261859e01b6b70ac35e96b0f3578090ddd36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941475
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>