[assembly] Move kernel into its own platform input bundle
Move the kernel out of the legacy input bundle, and into its own
platform AIB, using a special template that can use the metadata-based
contract with the kernel portion of the build to locate the path to the
file.

This also changes assembled_system() to create a 'custom kernel AIB' if
a non-standard kernel image is passed to the assembled_system() GN
template.  A new option is added to product assembly to receive the path
to that AIB and use it in lieu of the standard kernel image in the
platform AIBs.

Bug: b/308686895
Change-Id: Ia42a7017f478de26636f04512e4d242a756f5c1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940438
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>