[kernel][arch] Introduce global support to the paging traits
This change adds support for global pages in the libarch paging traits.
This will be used in the eventual mapping of the kernel.

Because it was low-hanging exercise and otherwise still planned for, we
also enable global mappings in the phys environment and update the
virtual address loading test to map its test ELF image as global. This
is accomplished by adopted the policy that AddressSpace::Map() will
always fix up the input settings to set global when mapping in the upper
address space.

Bug: 91187
Change-Id: I3a62f102cb20debb12f92686b0610e0268804aed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900444
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>