[fidl][cpp] Use --assume-filename to speed up Clang-Format
Clang-Format uses the filename (specifically the extension) to determine
the language of the input. When reading the input from standard input,
it uses auto-detection, which tends to be expensive. We can instead
override the language using the --assume-filename option which in my
experiments reduces the runtime by about 20%. Note that the filename in
this case is irrelevant, what matters is the extension.

Change-Id: I44d6b434e0d9bdf275f4d6c1bce8042a046a18be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941660
Fuchsia-Auto-Submit: Petr Hosek <phosek@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>