[ffx] Fix ssh logic for unresponsive target
The ssh logic was incorrect -- if the ssh connection timed out, the
stdout-parsing function would complete, but we'd then fall through to a
non-timing-out stderr-parsing function, which meant that we might wait
indefinitely on a hanging connection (as would sometimes show up with
targets that had just rebooted).

Along with this fix, we also eliminate the repetitive checking
of the ssh logging option, fix the nested timeout code in
parse_ssh_connection(), and make the stderr-parsing handle more
messages that might appear before the desired error message.

Fixes: b/307600923
Change-Id: I4e80c13e8582ab20714edd0193785877c691af55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941593
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>