[bt][a2dp] Suspend stream if permit gets taken during setup
The starting of an A2DP stream is a multi-step asynchronous operation.
In some cases during the setup of A2DP audio, HFP audio can be requested
and will take priority by seizing A2DP's streaming permits. Update A2DP
to be resilient to this case by checking for the validity of the permit
before starting the local A2DP stream. If the permit is unavailable,
don't start the local stream and issue a subsequent Suspend to
synchroize the streaming state with the remote peer.

Add a unit test which reproduces this case and fails in the current
implementation.

Bug: b/306661742
Test: fx test bt-a2dp-lib-test

Change-Id: I326e1e5af29e8399fac65ad303d52def716cd966
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/939982
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>